script that updates cloud.yaml
The two main changes are that IPs for masquerades are fetched and stored in the
configuration files, and that cloud.yaml is now considered authoritative.  The
cert files were very inconvenient and it made little sense to extend them to
somehow contain IPs too.  To make cloud.yaml the single point of control, the
Jinja template that produces it, and where some data was authoritatively
stored, was ditched too.  cloud.yaml is now generated by using only the yaml
library.

The changes to actually use these IPs to connect to masquerades will be pushed
in a separate PR, since those have the cloud.yaml from this PR as
a prerequisite.