first shot at using the IP of masquerades instead of the domain
It seems if you dial directly to an IP Cloudflare will give you the
cert of the target site (i.e. getiantem.org), not the masquerade. Since IIUC
the cert is exchanged in the unencrypted part of the TLS handshake, this is
a fingerprintability issue.

I'm committing this so others can experiment with it.