Make TLS cipher suite order more closely match Go TLS library
TLS fingerprinting shows that there are a few differences between
Lantern’s fingerprint and the fingerprint generated by the standard Go
TLS library.  One of these relates to the order of the cipher suites.
Here, we reorder cipher suites to match the standard Go fingerprint on
desktop, where prioritizing AES ciphers has less of a performance
concern than on mobile.