Fix intermittently failing test
This test would fail if there is no `config/global.yaml` file, which
may happen the first time running the tests.  The test creates a
`global.yaml` file during its run.

With the `global.yaml` file, `InitWithURLs` will _not_ attempt to fetch
the embedded config, relying on that local file instead.  Else, it will
try to fetch the embedded config, which may be missing the poll
interval properties, causing the test to rely on the defaults which are
far longer than the test expects.