Fix intermittent `TestInitWithURLs` test failure
This test was intermittently failing because `TestInit` was fetching a global config from a remote server and storing it at `global.yaml`, which the `TestInitWithURLs` test also pulls from.  The remote config had poll intervals set to much longer durations, which caused this test to fail.  Now, we simply delete `global.yaml` once `TestInit` has completed and allow it to be regenerated in `TestInitWithURLs`.