Try to get the metainfo from S3 even if it's not the first time the infohash is seen
I found this while debugging. If you get the link wrong, or if there is an error getting the metainfo any time, there will be not be another attempt. For Replica objects on s3, this can be fatal because the s3 BitTorrent peer doesn't support metadata transfers. It could be worth wrapping this attempt in a singleflight down the track.