Do not allow port 25/25-TLS to proxy in SOCKS mode
This is too operationally risky. 

If it is not already being abused, this will be abused one day can cause our proxies to be null routed or the account with the provider to be put at risk.

Most cloud providers already filter this off for you anyway, so it's not like this is doing that much in the first place I suspect.