Attempt to smooth out requests to config server
Currently, we let the server dictate how long we should sleep, so that we can
kick requests down the road at times of high load. When config-server has high
load and takes more than 3s to respond, it instructs the client to sleep for a
random amount of time between five and ten minutes.

Since we sleep after the request completes and ignore the time the request
itself took, and because config-server finishes batches of requests all at once
after periods of slowness, this creates a natural synchronization; requests
finish around the same time, so the next one starts around the same time.
Unfortunately, the ~5m period coincides with a 5m periodic slowness on the
config-server side (because of Redis BGSAVE), which creates a kind of latency
superresonance.

The best way to avoid any sort of sawtoothiness with requests is to just ignore
everything and sleep for a random amount of time in between, so that's the
approach here. In particular, we ignore the request end time, and sleep based
on the request start time.