Improve certificate pinning
We pin TLS certificates when connecting to HTTPS proxies. This is because
our proxies use self-signed certificates, communicated out-of-band to clients
via the proxy config. We currently do this manually, through a byte-wise
comparison of the configured certificate and the certificate the server
provides.

One issue with this is that it defeats our "hello rolling" logic, which aims to
rotate to a new TLS ClientHello if anything goes wrong in the TLS handshake
with the proxy. If certificate verification happens outside of the handshake,
the hello-rolling logic will not trigger.

This causes a bug when clients try to connect to a session resumption proxy. We
use a ClientHello modeled after the user's default web browser. If the user's
browser does not support session tickets, the client's request will be
"reflected" to the proxy's masquerade origin. The client will realize something
is amiss when it checks the certificate, but because the check occurs outside
of the handshake, the client will not fallback to other (more reliable)
ClientHellos. Instead, the client will continue to use the bad ClientHello and
forever fail to establish a connection to the proxy.

The solution is to move to a more standard method of certificate verification.
We configure the proxy certificate as the root CA and let crypto/tls handle
verification as part of the handshake.